domainSchemas -> remove stripEmptyDataResolver code: move to separate file || utils -> new util stripEmptyDataResolver: exports a custom react hook form resolver that removes from the input the attributes with empty value and performs zod validation
